refactor: reorganize security-engineer as subagent package
- Move from plugins/ to subagents/ directory structure
- Change package name from @hugsylabs/plugin-security-engineer to @hugsylabs/subagent-security-engineer
- Export subagent configuration for use in .hugsyrc
- Update documentation for subagent usage pattern
