+++ .changeset/security-engineer-subagent.md
@@ -1,0 +1,10 @@
+---
+'@hugsylabs/subagent-security-engineer': minor
+---
+
+Add new security-engineer subagent package
+
+- Introduces a specialized security engineering subagent with deep expertise in DevSecOps, cloud security, and compliance frameworks
+- Provides comprehensive security analysis, vulnerability management, and incident response capabilities
+- Includes tools for infrastructure hardening, compliance automation, and zero-trust architecture implementation
+- Adapted from VoltAgent/awesome-claude-code-subagents (MIT License)
